Fix build error in cluster code.
git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk/qpid@691456 13f79535-47bb-0310-9956-ffa450edef68

+++ cpp/src/qpid/cluster/Connection.h
@@ -61,7 +61,7 @@ class Connection :
     // self-delivery of intercepted extension points.
     void deliver(framing::AMQFrame& f);
     void deliverClose();
-    void deliverDoOutput(size_t requested);
+    void deliverDoOutput(uint32_t requested);
 
     void codecDeleted();
